Move load_afq_data from insight.py to datasets.py
